Treeview complete
Also moved some things around to track with evolving app structure.
