+++ spiffworkflow-frontend/src/a-spiffui-v2/views/app/sidemenu/README.txt
@@ -1,0 +1,4 @@
+6/1/2024 This is the old side menu, which may still have some utility. 
+
+Leaving it in the code for now, if it becomes apparent we won't need it, 
+remove it.
